Move test to test_ceph_default_values_check.py
